[APPC-1301] Refactored BdsGamification Repository to create abstraction
